fix: transfer sign check in middleware
