+++ internal/handlers/updates.go
@@ -6,12 +6,10 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/ilya372317/must-have-metrics/internal/config"
 	"github.com/ilya372317/must-have-metrics/internal/dto"
 	"github.com/ilya372317/must-have-metrics/internal/logger"
 	"github.com/ilya372317/must-have-metrics/internal/server/entity"
 	"github.com/ilya372317/must-have-metrics/internal/server/service"
-	"github.com/ilya372317/must-have-metrics/internal/signature"
 )
 
 type BulkUpdateStorage interface {
@@ -23,18 +21,9 @@ type BulkUpdateStorage interface {
 	Update(ctx context.Context, name string, alert entity.Alert) error
 }
 
-func BulkUpdate(storage BulkUpdateStorage, serverConfig *config.ServerConfig) http.HandlerFunc {
+func BulkUpdate(storage BulkUpdateStorage) http.HandlerFunc {
 	return func(writer http.ResponseWriter, request *http.Request) {
 		writer.Header().Set("content-type", "application/json")
-		isSignCorrect, err := signature.IsCorrectSigned(serverConfig, request)
-		if err != nil {
-			http.Error(writer, fmt.Sprintf("failed check sign: %v", err), http.StatusInternalServerError)
-			return
-		}
-		if !isSignCorrect {
-			http.Error(writer, "invalid sign", http.StatusBadRequest)
-			return
-		}
 		metricsList, err := dto.NewMetricsListDTOFromRequest(request)
 		if err != nil {
 			http.Error(writer, fmt.Sprintf("failed create metricsList dto: %v", err), http.StatusBadRequest)
@@ -65,7 +54,6 @@ func BulkUpdate(storage BulkUpdateStorage, serverConfig *config.ServerConfig) ht
 			logger.Log.Warn(err)
 			return
 		}
-		signature.SetSign(writer, serverConfig, response)
 		if _, err = writer.Write(response); err != nil {
 			logger.Log.Warn(err)
 		}
